+++ code/AI/ai_lists/analyse/heuristiques.h
@@ -2,6 +2,6 @@
 #define HEURISTIQUE
 
 #include "../../../fundamental_functions/game_functions/Source/source_header.h"
-float heuristique_miam_trivial(Game *g);
+double heuristique_miam_trivial(Game *g);
 
 #endif
