+++ code/AI/alpha_beta_prunning/esperance_calc/esperance_calc.h
@@ -5,10 +5,10 @@
 #include "../../move_listing/move_listing.h"
 #include "../../ai_lists/ai_struct.h"
 
-float esperanceAlphaBetaPrunning(float (*f)(alphaBetaArg), alphaBetaArg alphaBetaArg, Move move, bool shouldFreeMove);
+double esperanceAlphaBetaPrunning(double (*f)(alphaBetaArg), alphaBetaArg alphaBetaArg, Move move, bool shouldFreeMove);
 //retourne l'esperance des appels reccusrsifs
 //de l'elagage alpha-beta
 
-float esperanceHeuristique(AI ai, Game* g, Move move);
+double esperanceHeuristique(AI ai, Game* g, Move move);
 //score moyen du coup move selon l'heuristique h
 #endif
