+++ code/fundamental_functions/game_functions/access_functions/access_functions.h
@@ -21,7 +21,7 @@ bool getFriendByInd(Game *g, int indActu, int indFriend, bool colorActu);
 coord_tab_t *friendTabToCoordTab(Game *g, int ind, bool color, int indSpecial);
 
 bool has_friend(Game *g, int ind, int color);
-float zero_fun(Game *g);
+double zero_fun(Game *g);
 Coord to_coord(int i, int j);
 Coord coord_init();
